requirements.txt
autopep8==1.4.4
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
idna==2.8
lxml==4.4.2
pycodestyle==2.5.0
pymongo==3.9.0
python-dotenv==0.10.3
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.7
